requirements.txt
bcrypt==3.2.2
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
mysqlclient==2.1.1
PyMySQL==1.0.2
requests==2.26.0
gunicorn==20.1.0
PyJWT==2.4.0